Growth School Project Claude/requirements.txt
flask==2.0.1
flask-cors==3.0.10
numpy>=1.24.0
pillow>=8.3.2
tensorflow-cpu>=2.6.0
scikit-learn>=0.24.2
python-dotenv>=0.19.0
google-generativeai 